This patch fixes a number of simple warnings at /W3 level
They mostly boils down to theses simple (and not too severe) errors:

 - Storing unsinged numbers in singed variables, then comparing them
 back to unsigned numbers
 - Declaring for loop index variables as signed when testing it against
 unsigned values (eg: size of a container)
 - Using undeclared funcions in C. Compiler will fallback to thinking of
 them as extern function returning `int`. This *happens* to work because
 the linker will find the correct function later.

 **Note**: This path doesn't replace the usage of "unsafe" C standard function
 with their "safe" (generally postfixed _s) counterpart. I'm planning to
 submit a separate PR for that ;-)

+++ src/wfdir.c
@@ -599,7 +599,7 @@ DirWndProc(
       UINT  cItems;
       LPWSTR szItem;
       WCHAR rgchMatch[MAXPATHLEN];
-      INT cchMatch;
+      SIZE_T cchMatch;
 
       if ((ch = LOWORD(wParam)) <= CHAR_SPACE || !GetWindowLongPtr(hwnd, GWL_HDTA))
          return(-1L);
